+++ src/main/java/com/hanhwa_tae/gulhan/user/me/query/controller/MyPageQueryController.java
@@ -1,12 +1,15 @@
 package com.hanhwa_tae.gulhan.user.me.query.controller;
 
 import com.hanhwa_tae.gulhan.auth.command.domain.aggregate.model.CustomUserDetail;
+import com.hanhwa_tae.gulhan.common.domain.TargetType;
 import com.hanhwa_tae.gulhan.common.dto.ApiResponse;
 import com.hanhwa_tae.gulhan.common.exception.BusinessException;
 import com.hanhwa_tae.gulhan.common.exception.ErrorCode;
 import com.hanhwa_tae.gulhan.user.me.query.dto.GoodsOrderHistoryDTO;
 import com.hanhwa_tae.gulhan.user.me.query.dto.PackageOrderHistoryDTO;
 import com.hanhwa_tae.gulhan.user.me.query.service.MyPageQueryService;
+import com.hanhwa_tae.gulhan.user.query.dto.response.UserCommentResponse;
+import com.hanhwa_tae.gulhan.user.query.dto.response.UserReviewResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -15,7 +18,7 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @RestController
-@RequestMapping("/api/v1/user/me")
+@RequestMapping("/api/v1/users/me")
 @RequiredArgsConstructor
 public class MyPageQueryController {
 
@@ -59,6 +62,25 @@ public class MyPageQueryController {
         Long userNo = user.getUserNo();
         myPageQueryService.confirmOrder(orderHistoryId);
         return ResponseEntity.ok(ApiResponse.success(null));
+    }
+
+    @GetMapping("/review")
+    public ResponseEntity<ApiResponse<UserReviewResponse>> getUserReview(
+            @AuthenticationPrincipal CustomUserDetail userDetail,
+            @RequestParam TargetType targetType
+    ){
+        UserReviewResponse response = myPageQueryService.getUserReview(userDetail, targetType);
+
+        return ResponseEntity.ok(ApiResponse.success(response));
+    }
+
+    @GetMapping("/comment")
+    public ResponseEntity<ApiResponse<UserCommentResponse>> getUserComment(
+            @AuthenticationPrincipal CustomUserDetail userDetail
+    ){
+        UserCommentResponse response = myPageQueryService.getUserComment(userDetail);
+
+        return ResponseEntity.ok(ApiResponse.success(response));
     }
 
 }
